[UX] fix: review round — word-box trim offset, preview OOM guard, selector allocs
- TtfWordSelect: advance the word box past trimmed leading punctuation so
  touch/highlight geometry matches the lookup text
- TextSettingsPreview: only commit preview runs when layoutPlainText
  returns BookStatus::Ok; a failed pass keeps the previous runs and key
- CrossPointSettings: read ttfFontPointSize through int so a malformed
  JSON value (e.g. 300) cannot truncate past the 8..72 range gate
- main: unconditional volatile read anchors bookStatusProbe against
  linker GC now that LOG_DBG compiles out on release builds
- EpubReaderActivity: makeUniqueNoThrow + OOM guard for both dictionary
  selector allocations (bare make_unique aborts under -fno-exceptions)
- progress stub: only writable handles sync on close; close propagates
  the save result (read handles never persist their copy)
- tests: CloseBookRejectsLateSave now exercises a FAILED close flush
  (closing_ + bookOpen_ true) and asserts the pre-close record; loader
  test asserts the second family's full face shape to catch leaks under
  its own path prefix

+++ src/CrossPointSettings.cpp
@@ -236,7 +236,9 @@ bool CrossPointSettings::fromJson(JsonVariantConst doc) {
   }
   const char* ttfFamily = doc["ttfFontFamilyName"] | "";
   copyToField(ttfFontFamilyName, ttfFamily, sizeof(ttfFontFamilyName));
-  const uint8_t storedTtfPointSize = doc["ttfFontPointSize"] | DEFAULT_TTF_FONT_POINT_SIZE;
+  // Read through int: a uint8_t pipe would truncate 300 to 44 and slip
+  // past the range gate below.
+  const int storedTtfPointSize = doc["ttfFontPointSize"] | static_cast<int>(DEFAULT_TTF_FONT_POINT_SIZE);
   if (storedTtfPointSize >= TTF_FONT_POINT_SIZE_MIN && storedTtfPointSize <= TTF_FONT_POINT_SIZE_MAX) {
     ttfFontPointSize = storedTtfPointSize;
   } else {
